refactor: extract chat context calculation
